home-backend/pyproject.toml
[tool.poetry]
name = "home"
version = "0.1.0"
description = ""
authors = ["alex1151212 <alex1151212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
motor = "^3.1.1"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^1.4.45"
python-multipart = "^0.0.5"
pyproject-toml = "^0.0.10"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
